backend/requirement.txt
flask
flask-cors
python-dotenv
requests
beautifulsoup4
geopy